fix: 修复部分 API 的已知错误
- 修复部分 API 的已知错误
- 修复各层级间的类型转换问题
- 修复评论类型中用户信息的数据结构问题

+++ domain/event/entities/event.go
@@ -9,7 +9,7 @@ type Event struct {
 	Name        string
 	Description string
 	IsDone      bool
-	SortId      uint32
+	SortId      uint16
 	CreatedAt   time.Time
 	UpdatedAt   time.Time
 }
